refactor: changing directory structure and renaming
